+++ FetchTreeChromeExt/contentScript.js
@@ -1,0 +1,3 @@
+import componentObj from '../FetchTreeNPMPkg/componentStore.js'
+
+console.log(componentObj);
