+++ FetchTreeChromeExt/devtools.js
@@ -1,3 +1,3 @@
 chrome.devtools.panels.create('React Fetch Tree', null,'panel.html', () => {
-  console.log('hi this ran');
+  console.log('hello world');
 });
